+++ src/utils/issue.ts
@@ -1,5 +1,5 @@
 import type { Issue, Status } from '~/types/issue';
-import { filterArray, reduceArray } from '~/utils/array';
+import { filterArray } from '~/utils/array';
 
 const filterIssueByStatus = ({ issues, status }: { issues: Issue[], status: Status }): Issue[] => {
   return filterArray({ arr: issues, filterFunc: (issue: Issue) => issue.status === status });
@@ -10,11 +10,3 @@ const setStatusToFilterIssue = (status: Status) => (issues: Issue[]) => filterIs
 export const filterOpenedIssues = setStatusToFilterIssue('open');
 
 export const filterClosedIssues = setStatusToFilterIssue('close');
-
-export const makeIssueTemplate = ({ issues, templateFunc }: { issues: Issue[], templateFunc: (issue: Issue) => string }): string => {
-  return reduceArray({
-    arr: issues,
-    initValue: '',
-    reducer: (template: string, issue: Issue) => `${template}${templateFunc(issue)}`
-  });
-};
